Remove http dependency
It may be nice but this dep has some native code inside which it compiles every time I install it. Since `ossy` is everywhere in dry-rb I spent ~1hr in total just waiting this particular dependency to install. This is a bit too much for a thing that does a single http call (in our case).